requirements.txt
alembic==0.9.5
altgraph==0.14
beautifulsoup4==4.5.1
blinker==1.4
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
future==0.16.0
itsdangerous==0.24
Jinja2==2.8.1
lamson==1.3.4
macholib==1.8
Mako==1.0.7
Markdown==2.6.7
markdown2==2.3.1
MarkupSafe==1.0
pefile==2017.9.3
Pillow==3.4.2
pygame==1.9.3
PyHook3==1.6.0
PyInstaller==3.3.dev0+g5dc9557
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
pywin32==221
requests==2.11.1
six==1.10.0
South==1.0.2
SQLAlchemy==1.1.14
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
visitor==0.1.3
Werkzeug==0.12.2
WMI==1.4.9
WTForms==2.1
